requirements.txt
numpy
pandas
scikit-learn
djangorestframework
django-cors-headers